Expand and refactor verification logic
Add a check that the SigningCertURL has the scheme 'https' in addition
to the URL domain check.

Perform verification in the main event handler logic, so that
notification messages are verified as well (not just subscription and
unsubscription messages).

Introduce a BaseMessage struct to hold common message fields, which also
allows us to define a common interface for the different message types,
simplifying the Verify() method.